requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-dotenv==1.4.2
django-tastypie==0.14.2
geoip2==2.9.0
idna==2.7
maxminddb==1.4.1
Pillow==5.3.0
psycopg2-binary==2.7.5
python-dateutil==2.7.4
python-mimeparse==1.6.0
pytz==2018.6
requests==2.20.0
six==1.11.0
urllib3==1.24
